refactor: tabs パッケージ内の整理とディレクトリ構成の再構築
tabs パッケージ配下のファイルを役割ごとにサブパッケージ（component, coordinator, dialog, model, registry, screen）へ移動し、整理しました。これに伴い、各ファイルのパッケージ宣言およびインポート文を更新しています。
