+++ app/src/main/java/com/websarva/wings/android/slevo/ui/tabs/dialog/UrlOpenDialog.kt
@@ -1,4 +1,4 @@
-package com.websarva.wings.android.slevo.ui.tabs
+package com.websarva.wings.android.slevo.ui.tabs.dialog
 
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.fillMaxWidth
