+++ app/src/main/java/com/websarva/wings/android/slevo/ui/tabs/model/BoardTabInfo.kt
@@ -1,4 +1,4 @@
-package com.websarva.wings.android.slevo.ui.tabs
+package com.websarva.wings.android.slevo.ui.tabs.model
 
 data class BoardTabInfo(
     val boardId: Long,
