+++ app/src/main/java/com/websarva/wings/android/slevo/ui/tabs/model/ThreadTabInfo.kt
@@ -1,4 +1,4 @@
-package com.websarva.wings.android.slevo.ui.tabs
+package com.websarva.wings.android.slevo.ui.tabs.model
 
 import com.websarva.wings.android.slevo.data.model.ThreadId
 import com.websarva.wings.android.slevo.data.model.threadKey
