+++ app/src/main/java/com/websarva/wings/android/slevo/ui/tabs/registry/TabViewModelRegistry.kt
@@ -1,4 +1,4 @@
-package com.websarva.wings.android.slevo.ui.tabs
+package com.websarva.wings.android.slevo.ui.tabs.registry
 
 import com.websarva.wings.android.slevo.ui.board.viewmodel.BoardViewModel
 import com.websarva.wings.android.slevo.ui.board.viewmodel.BoardViewModelFactory
