+++ app/src/test/java/com/websarva/wings/android/slevo/ui/tabs/BoardTabsCoordinatorTest.kt
@@ -2,8 +2,9 @@ package com.websarva.wings.android.slevo.ui.tabs
 
 import com.websarva.wings.android.slevo.data.repository.BookmarkBoardRepository
 import com.websarva.wings.android.slevo.data.repository.TabsRepository
-import com.websarva.wings.android.slevo.ui.navigation.AppRoute
-import io.mockk.coVerify
+import com.websarva.wings.android.slevo.ui.tabs.coordinator.BoardTabsCoordinator
+import com.websarva.wings.android.slevo.ui.tabs.model.BoardTabInfo
+import com.websarva.wings.android.slevo.ui.tabs.registry.TabViewModelRegistry
 import io.mockk.mockk
 import org.junit.Assert.assertEquals
 import org.junit.Test
