+++ app/src/test/java/com/websarva/wings/android/slevo/ui/thread/viewmodel/ThreadTabCoordinatorTest.kt
@@ -4,7 +4,7 @@ import com.websarva.wings.android.slevo.data.datasource.local.entity.ThreadReadS
 import com.websarva.wings.android.slevo.data.model.ThreadId
 import com.websarva.wings.android.slevo.data.repository.TabsRepository
 import com.websarva.wings.android.slevo.data.repository.ThreadReadStateRepository
-import com.websarva.wings.android.slevo.ui.tabs.ThreadTabInfo
+import com.websarva.wings.android.slevo.ui.tabs.model.ThreadTabInfo
 import io.mockk.Called
 import io.mockk.coVerify
 import io.mockk.every
